SHOWROOMの一部のファイル名を変更
