+++ ShowRoomSitePlugin/IShowRoomSiteOptions.cs
