+++ ShowRoomSitePlugin/Message/ShowRoomConnected.cs
