+++ ShowRoomSitePlugin/ShowRoomOptionsPanel.xaml
